+++ src/main/java/wooteco/subway/line/dto/LineResponse.java
@@ -1,7 +1,7 @@
-package wooteco.subway.line;
+package wooteco.subway.line.dto;
 
 import java.util.List;
-import wooteco.subway.station.StationResponse;
+import wooteco.subway.station.dto.StationResponse;
 
 public class LineResponse {
