+++ src/main/java/wooteco/subway/station/dao/StationDao.java
@@ -1,6 +1,7 @@
-package wooteco.subway.station;
+package wooteco.subway.station.dao;
 
 import java.util.List;
+import wooteco.subway.station.Station;
 
 public interface StationDao {
