+++ src/main/java/wooteco/subway/station/dto/StationDto.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station;
+package wooteco.subway.station.dto;
 
 public class StationDto {
