+++ src/test/java/wooteco/subway/line/LineDaoCacheTest.java
@@ -2,7 +2,6 @@ package wooteco.subway.line;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
@@ -10,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import wooteco.subway.exception.DuplicatedLineNameException;
 import wooteco.subway.exception.VoidLineException;
+import wooteco.subway.line.dao.LineDaoCache;
 
 class LineDaoCacheTest {
