+++ src/test/java/wooteco/subway/line/LineResponseTest.java
@@ -6,7 +6,8 @@ import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import wooteco.subway.station.StationResponse;
+import wooteco.subway.line.dto.LineResponse;
+import wooteco.subway.station.dto.StationResponse;
 
 class LineResponseTest {
