+++ src/test/java/wooteco/subway/line/LineServiceTest.java
@@ -13,6 +13,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import wooteco.subway.assembler.Assembler;
 import wooteco.subway.exception.VoidLineException;
+import wooteco.subway.line.dao.LineDaoCache;
+import wooteco.subway.line.dto.LineDto;
 
 public class LineServiceTest {
