+++ src/test/java/wooteco/subway/station/StationDaoCacheTest.java
@@ -7,6 +7,7 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.station.dao.StationDaoCache;
 
 class StationDaoCacheTest {
