+++ src/test/java/wooteco/subway/station/StationRequestTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.station.dto.StationRequest;
 
 class StationRequestTest {
