+++ src/test/java/wooteco/subway/station/StationResponseTest.java
@@ -1,10 +1,10 @@
 package wooteco.subway.station;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.station.dto.StationResponse;
 
 class StationResponseTest {
