+++ src/test/java/wooteco/subway/station/StationServiceTest.java
@@ -10,6 +10,8 @@ import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.station.dao.StationDaoCache;
+import wooteco.subway.station.dto.StationDto;
 
 class StationServiceTest {
